requirements.txt
absl-py==1.4.0
altair==4.2.2
altair-data-server==0.4.1
altair-viewer==0.4.0
attrs==22.2.0
blinker==1.6
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
filelock==3.10.7
Fiona==1.9.2
fonttools==4.39.3
gadm==0.0.3
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.1
google-auth-oauthlib==1.0.0
grpcio==1.53.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mercantile==1.2.1
mpmath==1.3.0
munch==2.5.0
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
portpicker==1.5.2
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
seaborn==0.12.2
semver==3.0.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.20.0
streamlit-image-comparison==0.0.4
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
tifffile==2023.3.21
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
types-requests==2.28.11.17
types-setuptools==67.6.0.7
types-urllib3==1.26.25.10
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
zipp==3.15.0
